ai-vision-assistant/backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
torch==2.0.1
torchvision==0.15.2
opencv-python==4.8.1.78
Pillow==10.0.1
pytesseract==0.3.10
gTTS==2.4.0
numpy==1.24.3
transformers==4.33.1
pyyaml==6.0.1
maskrcnn-benchmark @ git+https://github.com/microsoft/GLIP.git
matplotlib==3.7.2
pycocotools==2.0.7
cityscapesscripts==2.2.2
scipy==1.11.2